Added JSON fields formats

+++ src/main/java/com/khaled/tms/Services/UsersServices.java
@@ -3,15 +3,19 @@ package com.khaled.tms.Services;
 import com.khaled.tms.Entity.UserEntity;
 import com.khaled.tms.Repo.UserRepo;
 import jakarta.persistence.EntityNotFoundException;
+import lombok.extern.slf4j.Slf4j;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
 
+@Slf4j
 @Service
+@Transactional
 public class UsersServices {
 
     private static final Logger log = LoggerFactory.getLogger(UsersServices.class);
@@ -51,14 +55,36 @@ public class UsersServices {
         return userRepo.save(uservar);
     }
 
-
+    //TODO: DELETE NOT WORKING
     public String deleteUser(String username) {
-//        if (!userRepo.existsByUsername(username)) {
+        username = username.toLowerCase();
+if (CheckIfUserExists(username)) {
+    userRepo.deleteByUsername(username);
+    log.info("User {} deleted", username);
+    return "User with username: " + username + " was deleted";
+}
+    throw new EntityNotFoundException("User with username " + username + " not found");
+
+}
+
+//    public String deleteUser(String username) {
+//        String normalized = username.trim().toLowerCase(); // sanitize
+//    log.info("Deleting userrrrr with id {}", normalized);
+//        Optional<UserEntity> userOpt = userRepo.findByUsername(normalized);
+//
+//        if (userOpt.isEmpty()) {
 //            throw new EntityNotFoundException("User with username " + username + " not found");
 //        }
-        userRepo.deleteByUsername(username);
-        log.info("User {} deleted", username);
-        return "User with username: " + username + " was deleted";
+//
+//        userRepo.delete(userOpt.get());
+//        log.info("User {} deleted", normalized);
+//        return "User with username: " + normalized + " was deleted";
+//    }
+
+
+    public boolean CheckIfUserExists(String username) {
+        log.info("Checking if user with username {} exists", username);
+        return userRepo.existsByUsername(username);
     }
 
     }
